test: Reintroduced IBM.Watson tests and moved AdditionalProperties tests into project
