+++ IBM.Watson.sln
@@ -59,7 +59,7 @@ Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "IBM.Watson.Assistant.v2.Int
 EndProject
 Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "IBM.Watson.Assistant.v1.IntegrationTests", "test\Assistant.v1.IntegrationTests\IBM.Watson.Assistant.v1.IntegrationTests.csproj", "{8204EB2C-938D-4C16-BE60-15DA66F9D70C}"
 EndProject
-Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "AdditionalPropertiesTest", "AdditionalPropertiesTest\AdditionalPropertiesTest.csproj", "{F114E55F-6A87-431C-813B-1ABFF8FE7569}"
+Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "IBM.Watson.IntegrationTests", "test\IBM.Watson.IntegrationTests\IBM.Watson.IntegrationTests.csproj", "{76EDAFFC-A536-40F3-B744-D894A5B28F46}"
 EndProject
 Global
 	GlobalSection(SolutionConfigurationPlatforms) = preSolution
@@ -127,10 +127,10 @@ Global
 		{8204EB2C-938D-4C16-BE60-15DA66F9D70C}.Debug|Any CPU.Build.0 = Debug|Any CPU
 		{8204EB2C-938D-4C16-BE60-15DA66F9D70C}.Release|Any CPU.ActiveCfg = Release|Any CPU
 		{8204EB2C-938D-4C16-BE60-15DA66F9D70C}.Release|Any CPU.Build.0 = Release|Any CPU
-		{F114E55F-6A87-431C-813B-1ABFF8FE7569}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
-		{F114E55F-6A87-431C-813B-1ABFF8FE7569}.Debug|Any CPU.Build.0 = Debug|Any CPU
-		{F114E55F-6A87-431C-813B-1ABFF8FE7569}.Release|Any CPU.ActiveCfg = Release|Any CPU
-		{F114E55F-6A87-431C-813B-1ABFF8FE7569}.Release|Any CPU.Build.0 = Release|Any CPU
+		{76EDAFFC-A536-40F3-B744-D894A5B28F46}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{76EDAFFC-A536-40F3-B744-D894A5B28F46}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{76EDAFFC-A536-40F3-B744-D894A5B28F46}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{76EDAFFC-A536-40F3-B744-D894A5B28F46}.Release|Any CPU.Build.0 = Release|Any CPU
 	EndGlobalSection
 	GlobalSection(SolutionProperties) = preSolution
 		HideSolutionNode = FALSE
